Queue processor moved to own file
